+++ src/api/windows/winapi64/kernel32/load_resource.rs
@@ -2,7 +2,7 @@ use crate::emu;
 
 pub fn LoadResource(emu: &mut emu::Emu) {
     let hModule = emu.regs().rcx;
-    let hResInfo = emu.regs().rdx as u64;
+    let hResInfo = emu.regs().rdx;
 
     log_red!(
         emu,
