+++ src/engine/instructions/fsubr.rs
@@ -10,7 +10,7 @@ pub fn execute(emu: &mut Emu, ins: &Instruction, instruction_sz: usize, _rep_ste
     let value0 = emu.get_operand_value(ins, 0, false).unwrap_or(0) as usize;
     let value1 = emu.get_operand_value(ins, 1, false).unwrap_or(0) as usize;
 
-    emu.fpu_mut().subr(value0 as usize, value1 as usize);
+    emu.fpu_mut().subr(value0, value1);
     emu.sync_fpu_ip();
     true
 }
